Add registries layout
