+++ lib/registries/addon/components/registries-layout/block/template.hbs
@@ -1,0 +1,3 @@
+<div ...attributes>
+    {{yield}}
+</div>
