+++ lib/registries/addon/components/registries-layout/styles.scss
@@ -1,0 +1,19 @@
+.Main {
+    min-height: 500px;
+    flex-grow: 1;
+}
+
+.RightGutter {
+    padding: 20px;
+    min-width: 300px;
+    width: 300px;
+}
+
+.ContentBackground {
+    display: flex;
+    flex-grow: 1;
+    justify-content: center;
+    background-image: url('img/bg-light.jpg');
+    background-repeat: no-repeat;
+    background-size: cover;
+}
