+++ lib/registries/addon/components/registries-layout/template.hbs
@@ -1,0 +1,46 @@
+{{yield (hash
+    heading=(component 'registries-layout/block')
+    sidenavGutterClosed=this.sidenavGutterClosed
+    toggleSidenav=(action this.toggleSidenav)
+    metadataGutterClosed=this.metadataGutterClosed
+    toggleMetadata=(action this.toggleMetadata)
+)}}
+
+<div local-class='ContentBackground'>
+    <Gutters
+        @leftMode={{this.sidenavGutterMode}}
+        @leftClosed={{this.sidenavGutterClosed}}
+        @rightMode={{this.metadataGutterMode}}
+        @rightClosed={{this.metadataGutterClosed}}
+        as |gutters|
+    >
+        {{#unless this.hideLeftGutter}}
+            <gutters.left-gutter>
+                {{yield (hash
+                    left=(component 'side-nav'
+                        onLinkClicked=(action this.toggleSidenav)
+                    )
+                )}}
+            </gutters.left-gutter>
+        {{/unless}}
+
+        <gutters.body local-class='Main'>
+            {{yield (hash
+                main=(component 'registries-layout/block')
+            )}}
+        </gutters.body>
+
+        {{#unless this.hideRightGutter}}
+            <gutters.right-gutter
+                local-class='RightGutter'
+                as |rightGutter|
+            >
+                {{yield (hash
+                    right=(component 'registries-layout/block')
+                    rightGutterMode=rightGutter.mode
+                    toggleMetadata=(action this.toggleMetadata)
+                )}}
+            </gutters.right-gutter>
+        {{/unless}}
+    </Gutters>
+</div>
