+++ lib/registries/addon/overview/-components/tombstone-page/styles.scss
@@ -1,8 +1,3 @@
-.TombstoneContainer {
-    flex-direction: column;
-    align-items: flex-start;
-}
-
 .Tombstone {
     background-color: $color-bg-white;
     margin: 30px 0;
