+++ lib/registries/addon/overview/-components/tombstone-page/template.hbs
@@ -1,27 +1,25 @@
-<Container
+<div
+    local-class='Tombstone'
     data-analytics-scope='Tombstone page'
-    local-class='TombstoneContainer'
     ...attributes
 >
-    <div local-class='Tombstone'>
-        <div local-class='TitleBlock'>
-            <div local-class='Title'>
-                <FaIcon
-                    local-class='Icon'
-                    @icon={{@icon}}
-                    @size=5
-                />
-                <h2 data-test-tombstone-title>
-                    {{@titleText}}
-                </h2>
-            </div>
-            <div local-class='SubtitleBlock'>
-                {{yield}}
-            </div>
+    <div local-class='TitleBlock'>
+        <div local-class='Title'>
+            <FaIcon
+                local-class='Icon'
+                @icon={{@icon}}
+                @size=5
+            />
+            <h2 data-test-tombstone-title>
+                {{@titleText}}
+            </h2>
         </div>
-        <RegistriesMetadata
-            local-class='MetadataBlock'
-            @registration={{this.registration}}
-        />
+        <div local-class='SubtitleBlock'>
+            {{yield}}
+        </div>
     </div>
-</Container>
+    <RegistriesMetadata
+        local-class='MetadataBlock'
+        @registration={{this.registration}}
+    />
+</div>
