+++ lib/registries/addon/overview/styles.scss
@@ -23,27 +23,14 @@
     align-items: center;
 }
 
+.MetadataTitle {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 0;
+}
+
 .OverviewBody {
-    min-height: 500px;
-    flex-grow: 1;
-    padding: 0 30px;
-}
-
-.SideNav {
-    height: 100%;
-    overflow-y: auto;
-    padding: 20px;
-}
-
-.SideMetadata {
-    padding: 20px;
-    min-width: 300px;
-    width: 300px;
-
-    .MetadataTitle {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-top: 0;
-    }
+    background: $white;
+    padding: 5px 30px 0;
 }
